fix: undo snackbar duration and entry name restoration
- Change action snackbar duration from Short to Long (~10s) so
  users have time to tap Undo before auto-dismiss
- Add entryName parameter to appendToHistory so restored entries
  keep their original name instead of resetting to blank

+++ composeApp/src/commonMain/kotlin/com/medqb/app/shared/domain/SnackbarMessage.kt
@@ -10,7 +10,7 @@ sealed interface SnackbarMessage {
     data class Action(
         override val message: String,
         val actionLabel: String,
-        val duration: SnackbarDuration = SnackbarDuration.Short,
+        val duration: SnackbarDuration = SnackbarDuration.Long,
         val onActionPerformed: () -> Unit,
     ) : SnackbarMessage
 }
